mini_lms_backend/apps/videos/urls.py
from rest_framework.routers import DefaultRouter
from .views import VideoViewSet
from django.urls import include, path

router = DefaultRouter()
router.register("", VideoViewSet, basename="videos")

urlpatterns = [path("", include(router.urls))]
