mini_lms_backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
dj-database-url
python-dotenv
gunicorn
django-cors-headers
cloudinary
django-cloudinary-storage
django-environ
